+++ src/raw_bmp.rs
@@ -8,6 +8,7 @@ use embedded_graphics::{
 use crate::{
     color_table::ColorTable,
     header::{Bpp, Header},
+    propagate,
     raw_iter::RawPixels,
     ChannelMasks, ParseError, RowOrder,
 };
@@ -39,20 +40,27 @@ impl<'a> RawBmp<'a> {
     ///
     /// The created object keeps a shared reference to the input and does not dynamically allocate
     /// memory.
-    pub fn from_slice(bytes: &'a [u8]) -> Result<Self, ParseError> {
-        let (_remaining, (header, color_table)) = Header::parse(bytes)?;
+    pub const fn from_slice(bytes: &'a [u8]) -> Result<Self, ParseError> {
+        let (_remaining, (header, color_table)) = propagate!(Header::parse(bytes));
 
-        let color_type = ColorType::from_header(&header)?;
+        let color_type = propagate!(ColorType::from_header(&header));
 
         // Believe what the bitmap tells us rather than multiplying width by
         // height by bits-per-pixel, because the image data might be compressed.
         let compressed_data_length = header.image_data_len as usize;
 
-        // The `get` is split into two calls to prevent an possible integer overflow.
-        let image_data = &bytes
-            .get(header.image_data_start..)
-            .and_then(|bytes| bytes.get(..compressed_data_length))
-            .ok_or(ParseError::UnexpectedEndOfFile)?;
+        if bytes.len() < header.image_data_start
+            || bytes.len() - header.image_data_start < compressed_data_length
+        {
+            return Err(ParseError::UnexpectedEndOfFile);
+        }
+        #[allow(unsafe_code)] // only way to resize a slice in const
+        let image_data = unsafe {
+            core::slice::from_raw_parts(
+                bytes.as_ptr().add(header.image_data_start),
+                compressed_data_length,
+            )
+        };
 
         Ok(Self {
             header,
@@ -160,7 +168,7 @@ pub enum ColorType {
 }
 
 impl ColorType {
-    pub(crate) fn from_header(header: &Header) -> Result<ColorType, ParseError> {
+    pub(crate) const fn from_header(header: &Header) -> Result<ColorType, ParseError> {
         Ok(match header.bpp {
             Bpp::Bits1 => ColorType::Index1,
             Bpp::Bits4 => ColorType::Index4,
@@ -182,7 +190,7 @@ impl ColorType {
             Bpp::Bits24 => ColorType::Rgb888,
             Bpp::Bits32 => {
                 if let Some(masks) = header.channel_masks {
-                    if masks == ChannelMasks::RGB888 {
+                    if let ChannelMasks::RGB888 = masks {
                         ColorType::Xrgb8888
                     } else {
                         return Err(ParseError::UnsupportedChannelMasks);